chore: moved string constants to partial class
