+++ Source/Cogworks.Essentials/Constants/StringConstants.cs
@@ -1,0 +1,6 @@
+﻿namespace Cogworks.Essentials.Constants
+{
+    public static partial class StringConstants
+    {
+    }
+}
