--- Source/Cogworks.Essentials/Constants/StringConstants/Separators.cs
@@ -1,18 +1,0 @@
-﻿namespace Cogworks.Essentials.Constants.StringConstants
-{
-    public static class Separators
-    {
-        public const string Comma = ",";
-        public const string Space = " ";
-        public const string Underscore = "_";
-        public const string CommaAndSpace = Comma + Space;
-        public const string Slash = "/";
-        public const string Backslash = @"\";
-        public const string Dot = ".";
-        public const string Semicolon = ";";
-        public const string Assign = ":=";
-        public const string Vertical = "|";
-        public const string Hyphen = "-";
-        public const string HtmlNewLine = "<br />";
-    }
-}
