+++ Source/Cogworks.Essentials/Extensions/EnumerableExtensions.cs
@@ -2,7 +2,7 @@
 using System.Collections;
 using System.Collections.Generic;
 using System.Linq;
-using Cogworks.Essentials.Constants.StringConstants;
+using Separators = Cogworks.Essentials.Constants.StringConstants.Separators;
 
 namespace Cogworks.Essentials.Extensions
 {
